backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
werkzeug==2.3.7
google-generativeai==0.3.2
requests==2.31.0
speechrecognition==3.10.0
pyttsx3==2.90
python-dotenv==1.0.0
gunicorn==21.2.0